Update reimbursementCalculator.py
Instead of manually imputing start and stop dates, the reimbursement calculator now reads a json file and calculates reimbursement rates based on data from the external json file. 
User must enter the name of a json file (reimbursement.json is included as an example) to calculate reimbursement rates. The json file must be in the same directory as the reimbursement calculator.